+++ jans-cli-tui/cli_tui/utils/utils.py
@@ -5,7 +5,7 @@ import prompt_toolkit
 
 from cli_style import style
 from wui_components.jans_drop_down import DropDownWidget
-from wui_components.jans_data_picker import DateSelectWidget
+from cli_tui.wui_components.jans_date_picker import DateSelectWidget
 from wui_components.jans_spinner import Spinner
